Allow object stream operations to retain original subclass (#122)
The new ServiceX Client needs to be able to pass state related to its subclass of ObjectStream from each func_adl statement on to the next. This change makes use of a deep copy of the target object and returns that from the func_adl verbs (Select, SelectMany, Where) that way at the end of the select we still have access to the original instance.