Type-Checking the Zip operator on Dictionary should not crash (#149)
* feat: Add test for dictionary with Zip key

Add a new test case to check that the type follows from a dictionary through a Zip operation. The test creates a lambda expression that selects the 'pt' attribute from a dictionary of jets, which are obtained from an event object. The test ensures that the expected exception is raised when an invalid key is used.

Refactor the test_type_based_replacement.py file to include the new test_dictionary_Zip_key() function.

* Fix up type and spelling errors

* Fix up more spelling and type checking errors

* Ignore `zip` when looking at dictionary attributes

* Fix up bad addition